+++ eslint.config.js
@@ -3,7 +3,7 @@ import config from '@biem/eslint-config';
 /** @type {import('eslint').Linter.Config[]} */
 export default [
 	...config,
-	{ ignores: ['src/template'] },
+	{ ignores: ['src/preset'] },
 	{
 		rules: {
 			'max-len': ['error', { code: 140 }],