+++ src/core/prompt/deploy/deployment-type.prompt.ts
@@ -6,13 +6,13 @@ import {
 	SimpleRenderer,
 } from 'listr2';
 
-import { CreateTemplateCtx } from '../../shared/ctx/create-template.ctx.js';
-import { DeploymentType } from '../../shared/enum/deployment-type.enum.js';
+import { DeployTemplateCtx } from '../../../shared/ctx/deploy-template.ctx.js';
+import { DeploymentType } from '../../../shared/enum/deployment-type.enum.js';
 
 export class DeploymentTypePrompt {
 	public async ask(
 		task: ListrTaskWrapper<
-			CreateTemplateCtx,
+			DeployTemplateCtx,
       typeof DefaultRenderer,
       typeof SimpleRenderer
 		>,