+++ src/core/prompt/deploy/workspace.prompt.ts
@@ -6,25 +6,25 @@ import {
 	SimpleRenderer,
 } from 'listr2';
 
-import { WorkspaceService } from '../../core/service/api/workspace.service.js';
-import { CreateTemplateCtx } from '../../shared/ctx/create-template.ctx.js';
-import { Workspace } from '../../shared/model/workspace.model.js';
+import { DeployTemplateCtx } from '../../../shared/ctx/deploy-template.ctx.js';
+import { Workspace } from '../../../shared/model/workspace.model.js';
+import { WorkspaceApiService } from '../../service/api/workspace.api.service.js';
 
 export class WorkspacePrompt {
-	private workspaceService: WorkspaceService;
+	private workspaceApiService: WorkspaceApiService;
 
-	constructor(workspaceService: WorkspaceService) {
-		this.workspaceService = workspaceService;
+	constructor(workspaceApiService: WorkspaceApiService) {
+		this.workspaceApiService = workspaceApiService;
 	}
 
 	public async ask(
 		task: ListrTaskWrapper<
-			CreateTemplateCtx,
+			DeployTemplateCtx,
       typeof DefaultRenderer,
       typeof SimpleRenderer
 		>,
 	): Promise<Workspace['id']> {
-		const { workspaceList } = await this.workspaceService.find();
+		const { workspaceList } = await this.workspaceApiService.find();
 
 		return task.prompt(ListrInquirerPromptAdapter).run(select, {
 			message: 'Select workspace',